feat: implement routes protection for photos searching
